feat(support): build real support hub, transparency page, and maintenance fund UI with local preview state
- replace placeholder support routes with real UI for /global_settings/support_home, /global_settings/support_transparency, and /global_settings/support_maintenance_fund
- move support screens into the global settings feature so support UX lives with the rest of the settings system
- add reusable SupportProgressCard for transparent monthly target, raised amount, remaining amount, supporter count, and status display
- make support state meaningfully react to current OPEN/CLOSED preview state from AppBootstrapController
- keep maintenance support optional and never tied to feature unlocks
- keep transparency page always visible and clearly explain the locked support rules
- make Support Home dynamically show maintenance-fund vs coffee path based on OPEN/CLOSED state
- make Maintenance Fund screen behave like a real UI-first funding surface while remaining honest that payment flow is not yet implemented
- keep local preview controls available so OPEN/CLOSED state can be tested during UI-first development
- update router so support routes now use real global settings screens instead of placeholders
- preserve existing route names, shell behavior, guest-first policy, and current architecture

+++ lib/features/plans/presentation/screens/plans_screens.dart
@@ -512,7 +512,7 @@ class PlansProgressScreen extends StatelessWidget {
     }).length;
     final int percent = totalDays == 0
         ? 0
-        : (((completedDays / totalDays) * 100).round().clamp(0, 100) as int);
+        : (((completedDays / totalDays) * 100).round().clamp(0, 100));
 
     return TabScreenScaffold(
       title: 'Progress',
